requirements.txt
matplotlib==3.4.3
neptune_client==0.13.1
numpy==1.21.4
Pillow==8.4.0
scikit_learn==1.0.1
torch==1.10.0
torchvision==0.11.1
